PASS-1: Pylint; mypy; fixes
