--- .markdownlint.json
@@ -1,5 +1,0 @@
-{
-    "default": true,
-    "MD024": false,
-    "MD013": false
-}
