+++ src/operator/cli.py
@@ -3,28 +3,23 @@ A kubernetes operator that syncs and decrypts secrets from Linux password store 
 """
 
 
-from __future__ import annotations
+from typing import Any
+from pathlib import Path
+from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
+from importlib import metadata
+from ipaddress import IPv4Address
+from kubernetes import client, config
+from src.operator.git import GitRepo
+from src.operator.utils import LogLevel
+from src.operator.secret import PassSecret
 
 import logging
 import sys
 import kopf
 import os
-import base64
-
-from typing import Any, Dict
-from kubernetes import client, config
-from pathlib import Path
-from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
-from importlib import metadata as meta
-from ipaddress import IPv4Address
-from dataclasses import dataclass, field
-
-from src.passoperator.git import GitRepo
-from src.passoperator.utils import LogLevel
-from src.passoperator.gpg import decrypt
 
 
-__version__ = meta.version('pass-operator')
+__version__ = metadata.version('pass-operator')
 
 log = logging.getLogger(__name__)
 pass_git_repo: GitRepo
@@ -32,174 +27,20 @@ config.load_incluster_config()
 
 
 # Environment variables to configure the operator's performance.
-OPERATOR_INTERVAL = int(os.getenv('OPERATOR_INTERVAL', 60))
-OPERATOR_INITIAL_DELAY = int(os.getenv('OPERATOR_INITIAL_DELAY', 3))
-OPERATOR_PRIORITY = int(os.getenv('OPERATOR_PRIORITY', 100))
+OPERATOR_INTERVAL = int(os.getenv('OPERATOR_INTERVAL', '60'))
+OPERATOR_INITIAL_DELAY = int(os.getenv('OPERATOR_INITIAL_DELAY', '3'))
+OPERATOR_PRIORITY = int(os.getenv('OPERATOR_PRIORITY', '100'))
 OPERATOR_NAMESPACE = os.getenv('OPERATOR_NAMESPACE', 'default')
 OPERATOR_POD_IP = IPv4Address(os.getenv('OPERATOR_POD_IP', '0.0.0.0'))
 
 # Environment variables to configure pass.
 PASS_BINARY = os.getenv('PASS_BINARY', '/usr/bin/pass')
 PASS_DIRECTORY = os.getenv('PASS_DIRECTORY', 'repo')
+PASS_GPG_PASSPHRASE = os.getenv('PASS_GPG_PASSPHRASE')
 PASS_GPG_KEY = os.getenv('PASS_GPG_KEY')
 PASS_GPG_KEY_ID = os.getenv('PASS_GPG_KEY_ID')
-PASS_GPG_PASSPHRASE = os.getenv('PASS_GPG_PASSPHRASE')
 PASS_GIT_URL = os.getenv('PASS_GIT_URL')
 PASS_GIT_BRANCH = os.getenv('PASS_GIT_BRANCH', 'main')
-
-
-@dataclass
-class ManagedSecret:
-    name: str
-    namespace: str ='default'
-    data: Dict[str, str] =dict()
-    stringData: Dict[str, str] =dict()
-    immutable: bool =False
-    secretType: str ='Opaque'
-    kind: str ='Secret'
-    apiGroup: str =''
-    apiVersion: str ='v1'
-
-    def __post_init__(self) -> None:
-        if not self.data and not self.stringData:
-            raise RuntimeError('ManagedSecret type expects at least one of \'data\', \'stringData\', to be set.')
-
-        # Propagate one field to the other, if only one or the other is set.
-        if self.data and not self.stringData:
-            self.stringData = {
-                key: base64.b64decode(value).rstrip().decode() for key, value in self.data.items()
-            }
-
-        if self.stringData and not self.data:
-            self.data = {
-                key: base64.b64encode(bytes(value.rstrip().encode('utf-8'))).decode() for key, value in self.stringData.items()
-            }
-
-    def to_client_dict(self) -> Dict:
-        """
-        Output this secret to a dictionary with keys that match the arguments of kubernetes.client.V1Secret, for convenience.
-        """
-        return {
-            'api_version': f'{self.apiGroup}/{self.apiVersion}' if self.apiGroup else self.apiVersion,
-            'kind': self.kind,
-            'metadata': {
-                'name': self.name,
-                'namespace': self.namespace,
-            },
-            'string_data': self.stringData,
-            'type': self.secretType,
-            'immutable': self.immutable
-        }
-
-
-@dataclass
-class PassSecret:
-    name: str
-    managedSecretName: str
-    encryptedData: Dict[str, str]
-    namespace: str ='default'
-    kind: str ='PassSecret'
-    apiGroup: str ='secrets.premiscale.com'
-    apiVersion: str ='v1alpha1'
-    annotations: Dict[str, str] =dict()
-    labels: Dict[str, str] =dict()
-
-    managedSecret: ManagedSecret = field(init=False)
-    managedSecretNamespace: str ='default'
-    managedSecretType: str ='Opaque'
-    managedSecretImmutable: bool =False
-
-    def __post_init__(self) -> None:
-        if (decryptedData := self.decrypt()) is None:
-            raise ValueError(f'Could not decrypt data on PassSecret {self.name}')
-
-        self.managedSecret = ManagedSecret(
-            name=self.managedSecretName,
-            namespace=self.managedSecretNamespace,
-            stringData=decryptedData,
-            immutable=self.managedSecretImmutable,
-            secretType=self.managedSecretType
-        )
-
-    def to_dict(self) -> Dict:
-        """
-        Output this object as a K8s manifest as JSON.
-        """
-        return {
-            'apiVersion': f'{self.apiGroup}/{self.apiVersion}',
-            'kind': self.kind,
-            'metadata': {
-                'name': self.name,
-                'namespace': self.namespace,
-                'annotations': self.annotations,
-                'labels': self.labels
-            },
-            'spec': {
-                'encryptedData': self.encryptedData,
-                'managedSecret': {
-                    'name': self.managedSecret.name,
-                    'namespace': self.managedSecret.namespace,
-                    'type': self.managedSecret.secretType,
-                    'immutable': self.managedSecret.immutable
-                }
-            }
-        }
-
-    @classmethod
-    def from_dict(cls, manifest: Dict) -> PassSecret | None:
-        """
-        Parse a k8s manifest into a PassSecret dataclass.
-        """
-        try:
-            if 'annotations' in manifest and len(manifest['annotations']):
-                annotations = manifest['annotations']
-            else:
-                annotations = dict()
-
-            if 'labels' in manifest and len(manifest['labels']):
-                labels = manifest['labels']
-            else:
-                labels = dict()
-
-            return cls(
-                name=manifest['metadata']['name'],
-                namespace=manifest['metadata']['namespace'],
-                encryptedData=manifest['spec']['encryptedData'],
-                labels=labels,
-                annotations=annotations,
-                # Parse out managed secret fields into arguments.
-                managedSecretName=manifest['spec']['managedSecret']['name'],
-                managedSecretNamespace=manifest['spec']['managedSecret']['namespace'],
-                managedSecretType=manifest['spec']['managedSecret']['type'],
-                managedSecretImmutable=manifest['spec']['managedSecret']['immutable']
-            )
-        except (KeyError, ValueError) as e:
-            log.error(f'Could not parse PassSecret into dataclass: {e}')
-            return None
-
-    def decrypt(self) -> Dict[str, str] | None:
-        """
-        Decrypt the contents of this PassSecret's paths and store them on an attribute
-
-        Returns:
-            Optional[Dict[str, str]]: a dictionary of data keys and decrypted paths' values. If decryption was not possible, None.
-        """
-        stringData = dict()
-
-        for secretKey in self.encryptedData:
-            secretPath = self.encryptedData[secretKey]
-
-            decryptedSecret = decrypt(
-                Path(f'~/.password-store/{PASS_DIRECTORY}/{secretPath}').expanduser(),
-                passphrase=PASS_GPG_PASSPHRASE
-            )
-
-            if decryptedSecret:
-                stringData[secretKey] = decryptedSecret
-            else:
-                return None
-
-        return stringData
 
 
 @kopf.on.startup()
@@ -217,7 +58,7 @@ def reconciliation(**kwargs) -> None:
     """
     Reconcile user-defined PassSecrets with the state of the cluster.
     """
-    log.info(f'Reconciling cluster state')
+    log.info('Reconciling cluster state')
     pass_git_repo.pull()
     check_gpg_id()
 
@@ -228,7 +69,7 @@ def reconciliation(**kwargs) -> None:
 
 
 @kopf.on.update('secrets.premiscale.com', 'v1alpha1', 'passsecret')
-def update(old: Any, new: Any, meta: kopf.Meta, **_: Any) -> None:
+def update(old: kopf.BodyEssence | Any, new: kopf.BodyEssence | Any, meta: kopf.Meta, **_: Any) -> None:
     """
     An update was received on the PassSecret object, so attempt to update the corresponding Secret.
 
@@ -367,7 +208,7 @@ def check_gpg_id(path: Path = Path(f'~/.password-store/{PASS_DIRECTORY}/.gpg-id'
         remove (bool): indicate whether or not to remove this file, should it exist.
     """
     if path.exists():
-        with open(path, mode='r') as gpg_id_f:
+        with open(path, mode='r', encoding='utf-8') as gpg_id_f:
             if gpg_id_f.read().rstrip() != PASS_GPG_KEY_ID:
                 log.error(f'PASS_GPG_KEY_ID ({PASS_GPG_KEY_ID}) does not equal .gpg-id contained in {path}')
                 sys.exit(1)
@@ -412,7 +253,7 @@ def main() -> None:
         sys.exit(0)
 
     if not PASS_GIT_URL:
-        log.error(f'Must provide a valid git URL (PASS_GIT_URL)')
+        log.error('Must provide a valid git URL (PASS_GIT_URL)')
         sys.exit(1)
 
     # Set up our global git repository object.
