+++ src/operator/git.py
@@ -3,13 +3,12 @@ Methods to interact minimally with a Git repository.
 """
 
 
-import sys
-import logging
-import os
-
 from pathlib import Path
 from typing import Union
 from git import Repo
+
+import logging
+import os
 
 
 log = logging.getLogger(__name__)
@@ -41,7 +40,7 @@ class GitRepo:
             loc (str): Local file path to clone to.
         """
         if self.cloned:
-            log.warn(f'Repository at URL {self.repo_url} has already been cloned to location "{self.clone_location}". Skipping')
+            log.warning(f'Repository at URL {self.repo_url} has already been cloned to location "{self.clone_location}". Skipping')
             return None
 
         self.repo = Repo.clone_from(
@@ -59,6 +58,8 @@ class GitRepo:
         self.cloned = True
 
         log.info(f'Successfully cloned repo {self.repo_url} to password store {self.clone_location}')
+
+        return None
 
     def pull(self) -> None:
         """
